refactor(scanner): extract native line/content filters to a sibling module
Move the pure, stateless structure-aware classification helpers
(_is_encrypted_value_line, _content_is_encrypted, _content_has_sops_markers,
_looks_like_code_member_access, _segment_is_word_like and their regexes) out of
native.py into scanner/_native_filters.py. native.py drops ~48 code lines below
its pre-PR baseline, clearing the CodeScene "Lines of Code in a Single File"
hotspot gate, and the line/content filters now form a cohesive, independently
testable unit. The helpers are re-exported from native.py so engine.py and the
scanner tests keep their existing import paths; the direct unit tests import them
from the new canonical module. Behavior is unchanged.
